backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
asgiref==3.9.1
attrs==25.3.0
blinker==1.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.7.0
greenlet==3.2.3
gunicorn==23.0.0
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.6.3
packaging==25.0
pillow==11.3.0
propcache==0.3.2
python-dotenv==1.1.1
requests==2.32.4
SQLAlchemy==2.0.41
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.35.0
Werkzeug==3.1.3
yarl==1.20.1
